Clean up lab-value based labelers -- rename 'labeling functions' and 'LFs' to be 'Labelers' to be more consistent in naming
