+++ src/piton/extractors/omop.py
@@ -180,7 +180,7 @@ class _ConceptTableConverter(CSVExtractor):
 
         if unit is not None:
             metadata["unit"] = unit
-
+            
         if unit_concept_id is not None:
             metadata["unit_concept_id"] = unit_concept_id
